added methods in NotificationDB to delete all notifications sent by a specified event, or delete all notifications sent to a specific user
